+++ package.json
@@ -9,7 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@teamhanko/hanko-elements": "^1.3.0",
+    "@teamhanko/hanko-elements": "^1.4.0",
     "next": "15.1.7",
     "react": "^19.0.0",
     "react-dom": "^19.0.0"
