+++ src/components/HankoAuth.tsx
@@ -2,8 +2,6 @@ import { useEffect, useCallback, useState } from "react";
 import { useRouter } from "next/navigation";
 import { register, Hanko } from "@teamhanko/hanko-elements";
 
-import './hanko-style.css'
-
 const hankoApi = process.env.NEXT_PUBLIC_HANKO_API_URL || '';
  
 export default function HankoAuth() {
