+++ src/hanko starter components/HankoStarterInfo.tsx
@@ -1,7 +1,5 @@
 import React from 'react'
 
-import './hanko-starter-style.css'
-
 const HankoStarterInfo = () => {
   return (
     <div className="starterInfo">
