+++ src/pages/_app.tsx
@@ -1,5 +1,8 @@
 import "@/styles/globals.css";
+import "@/styles/hanko-style.css";
+import "@/styles/hanko-starter-style.css";
 import type { AppProps } from "next/app";
+
 
 
 export default function App({ Component, pageProps }: AppProps) {
