+++ src/styles/hanko-style.css
@@ -22,31 +22,3 @@
     --background-color: rgba(0, 0, 0, 0.32);
   }
 }  
-
-hanko-auth{
-  /* Font Styles */
-  --font-weight: 400;
-  --font-size:1.7vh;
-
-  /* Item Styles */
-  --item-height: 60px;
-  --item-margin: 0.5rem 0;
-
-  /* Container Styles */
-  --container-padding: 50px;
-  --container-max-width:700px;
-}
-
-hanko-profile{
-
-  /* Font Styles */
-  --font-size:1.5vh;
-
-  /* Container Styles */
-  --container-padding: 10px;
-  --container-max-width:400px;
-
-  /* Item Styles */
-  --item-height: 30px;
-
-}
